requirements.txt
flask>=2.3.0
anthropic>=0.25.0
werkzeug>=2.3.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
